CCCP-227, refactor: move periodic tasks to rbclib/periodic
